refactor(application): update application error imports to base/ path
Redirect all application-layer error classes from deleted flat files
to the new errors.base subpackage.
