+++ src/forging_blocks/application/errors/event_store_error.py
@@ -1,12 +1,10 @@
 """Event store error type for application-level operations."""
 
-from collections.abc import Mapping
-
+from forging_blocks.foundation.errors.base.error import Error
 from forging_blocks.foundation.errors.core import ErrorMessage
-from forging_blocks.foundation.errors.error import Error
 
 
-class EventStoreError[MetadataType: Mapping[str, object] = dict[str, object]](Error[MetadataType]):
+class EventStoreError[MetadataValueType = object](Error[MetadataValueType]):
     """Base error for event store operations."""
 
     def __init__(self, message: str) -> None:
