+++ src/forging_blocks/application/errors/unit_of_work_error.py
@@ -1,9 +1,7 @@
-from collections.abc import Mapping
-
-from forging_blocks.foundation.errors.error import Error
+from forging_blocks.foundation.errors.base.error import Error
 
 
-class UnitOfWorkError[MetadataType: Mapping[str, object] = dict[str, object]](Error[MetadataType]):
+class UnitOfWorkError[MetadataValueType = object](Error[MetadataValueType]):
     """Error raised when a Unit of Work operation fails."""
 
     pass
